Conformed Order.Payment model to PayPalPaymentRepresentable protocol
