+++ Sources/App/Models/Order+Payment.swift
@@ -10,6 +10,9 @@ extension Order {
         let orderID: Order.ID
         var externalID: String?
         var currency: String
+        var payee: String?
+        var successRedirect: String?
+        var failureRedirect: String?
         
         var paid: Int
         var refunded: Int
@@ -44,90 +47,3 @@ extension Order.Payment: Migration {
 extension Order.Payment: Content {}
 extension Order.Payment: Parameter {}
 extension Order.Payment: MySQLModel {}
-
-extension Order.Payment {
-    func paypal(on conn: DatabaseConnectable) -> Future<PayPal.Payment> {
-        let shipping = Address.query(on: conn).filter(\.orderID == self.orderID).filter(\.shipping == true).first()
-        let items = Item.query(on: conn).filter(\.orderID == self.orderID).all()
-        let order = Order.query(on: conn).filter(\.id == self.orderID).first()
-        
-        
-        return map(shipping, items, order) { shipping, items, order -> PayPal.Payment in
-            let address: PayPal.Address?
-            let recipient = order?.firstname + order?.lastname
-            if let street = shipping?.street, let city = shipping?.city, let country = shipping?.country, let zip = shipping?.zip {
-                address = try PayPal.Address(
-                    recipientName: recipient,
-                    defaultAddress: false,
-                    line1: street,
-                    line2: shipping?.street2,
-                    city: city,
-                    state: shipping?.state,
-                    countryCode: country,
-                    postalCode: zip,
-                    phone: order?.phone,
-                    type: nil
-                )
-            } else {
-                address = nil
-            }
-            
-            let listItems = try items.map { item in
-                return try PayPal.Payment.Item(
-                    quantity: String(describing: item.quantity),
-                    price: String(describing: item.price),
-                    currency: Currency(code: self.currency) ?? .usd,
-                    sku: item.sku,
-                    name: item.name,
-                    description: item.description,
-                    tax: String(describing: item.tax)
-                )
-            }
-            let list = try PayPal.Payment.ItemList(items: listItems, address: address, phoneNumber: nil)
-            
-            
-            let subtotal = items.map { item in item.price * item.quantity }.reduce(0, +)
-            let tax = items.map { item in item.tax }.reduce(0, +)
-            
-            let details = try DetailedAmount.Detail(
-                subtotal: String(describing: subtotal),
-                shipping: String(describing: self.shipping),
-                tax: String(describing: tax),
-                handlingFee: String(describing: self.handling),
-                shippingDiscount: String(describing: self.shippingDiscount),
-                insurance: String(describing: self.insurence),
-                giftWrap: String(describing: self.giftWrap)
-            )
-            
-            let total = subtotal + tax
-            let amount = try DetailedAmount(
-                currency: Currency(code: self.currency) ?? .usd,
-                total: String(describing: total),
-                details: details
-            )
-            
-            let transaction = try PayPal.Payment.Transaction(
-                amount: amount,
-                payee: Payee(email: "placeholder@example.com", merchant: nil, metadata: nil),
-                description: nil,
-                payeeNote: nil,
-                custom: nil,
-                invoice: nil,
-                softDescriptor: nil,
-                payment: .instantFunding,
-                itemList: list,
-                notify: nil
-            )
-            
-            return try PayPal.Payment(
-                intent: .sale,
-                payer: PaymentPayer(method: .paypal, funding: nil, info: nil),
-                context: nil,
-                transactions: [transaction],
-                experience: nil,
-                payerNote: nil,
-                redirects: Redirects(return: "https://placeholder.com/success", cancel: "https://placeholder.com/fail")
-            )
-        }
-    }
-}
